project: Reorder files and classes structure
